+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog
 1.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add IContact and IPerson behaviors in addition to the IAddress behavior.
+  [thet]
+
+- PEP 8.
+  [thet]
 
 
 1.0 (2014-04-30)